Add more values for plugins + improve booking change

+++ action-logger.php
@@ -64,11 +64,13 @@
 
                 // EM actions
                 add_action( 'em_bookings_deleted',   array( $this, 'al_log_registration_delete'), 10, 2 );
-                add_action( 'em_booking_save',       array( $this, 'al_log_registration_cancel_reject'), 10, 2 );
+                add_action( 'em_booking_save',       array( $this, 'al_log_registration_chance'), 10, 2 );
 
                 // Shortcode
                 add_shortcode( 'actionlogger',  array( $this, 'al_register_shortcode_logger' ) );
-
+    
+                // $this->al_set_default_values();
+                
             }
 
             // @TODO: restrict to roles
@@ -89,7 +91,7 @@
              */
             public function al_plugin_activation() {
                 $this->al_prepare_log_table();
-                $this->al_store_default_values();
+                $this->al_set_default_values();
             }
 
             /**
@@ -102,8 +104,10 @@
             public function al_plugin_deactivation() {
 
                 $available_options = get_option( 'al_available_log_actions' );
-                foreach( $available_options as $option ) {
-                    delete_option( 'al_' . $option[ 'action_name' ] );
+                if ( false != $available_options ) {
+                    foreach( $available_options as $option ) {
+                        delete_option( 'al_' . $option[ 'action_name' ] );
+                    }
                 }
                 delete_option( 'al_available_log_actions' );
                 delete_option( 'al_log_user_role' );
@@ -178,7 +182,7 @@
             /**
              * Here we built a simple array of available log actions and store them in an option value.
              */
-            public function al_store_default_values() {
+            public function al_set_default_values() {
 
                 $available_options = get_option( 'al_available_log_actions' );
                 // $available_options = false;
@@ -189,73 +193,107 @@
                             'action_generator'   => 'WordPress',
                             'action_title'       => 'Create user',
                             'action_description' => esc_html( __( 'Logs when a new user is created in WordPress.', 'action-logger' ) ),
+                            'default_value'      => 1,
                         ),
                         array(
                             'action_name'        => 'wp_user_change',
                             'action_generator'   => 'WordPress',
                             'action_title'       => 'Change user',
                             'action_description' => esc_html( __( 'Logs when a user is changed (by another user) in WordPress.', 'action-logger' ) ),
+                            'default_value'      => 1,
                         ),
                         array(
                             'action_name'        => 'wp_user_delete',
                             'action_generator'   => 'WordPress',
                             'action_title'       => 'Delete user',
                             'action_description' => esc_html( __( 'Logs when a user is deleted in WordPress', 'action-logger' ) ),
+                            'default_value'      => 1,
                         ),
                     );
+                    
                     $user_options      = array(
                         array(
                             'action_name'        => 'user_visit_visitor',
                             'action_generator'   => 'WordPress',
                             'action_title'       => 'User visit (visitor)',
                             'action_description' => esc_html( __( 'Logs when a registered user visits a post/page with the shortcode on it.', 'action-logger' ) ),
+                            'default_value'      => 1,
                         ),
                         array(
                             'action_name'        => 'user_visit_registered',
                             'action_generator'   => 'WordPress',
                             'action_title'       => 'User visit (registered)',
                             'action_description' => esc_html( __( 'Logs when a visitor visits a post/page with the shortcode on it.', 'action-logger' ) ),
+                            'default_value'      => 1,
                         ),
                     );
                     $available_options = array_merge( $available_options, $user_options );
-                    if ( class_exists( 'RankingsImport' ) ) {
-                        $csvi_options      = array(
-                            array(
-                                'action_name'        => 'csvi_file_uploaded',
-                                'action_generator'   => 'CSV Importer',
-                                'action_title'       => 'CSV file uploaded',
-                                'action_description' => esc_html( __( 'Logs when a csv file is uploaded', 'action-logger' ) ),
-                            ),
-                        );
-                        $available_options = array_merge( $available_options, $csvi_options );
-                    }
-	                $available_options = array_merge( $available_options, $user_options );
+    
+                    // add option for events manager
+                    $em_options        = array(
+                        array(
+                            'action_name'        => 'em_booking_approved',
+                            'action_generator'   => 'Events Manager',
+                            'action_title'       => 'Booking approved',
+                            'action_description' => esc_html( __( 'Logs when a booking is approved', 'action-logger' ) ),
+                            'default_value'      => 0,
+                        ),
+                        array(
+                            'action_name'        => 'em_booking_canceled',
+                            'action_generator'   => 'Events Manager',
+                            'action_title'       => 'Booking changed',
+                            'action_description' => esc_html( __( 'Logs when a booking is canceled', 'action-logger' ) ),
+                            'default_value'      => 0,
+                        ),
+                        array(
+                            'action_name'        => 'em_booking_rejected',
+                            'action_generator'   => 'Events Manager',
+                            'action_title'       => 'Booking canceled/rejected',
+                            'action_description' => esc_html( __( 'Logs when a booking is rejected', 'action-logger' ) ),
+                            'default_value'      => 0,
+                        ),
+                        array(
+                            'action_name'        => 'em_booking_deleted',
+                            'action_generator'   => 'Events Manager',
+                            'action_title'       => 'Booking delete',
+                            'action_description' => esc_html( __( 'Logs when a booking is deleted', 'action-logger' ) ),
+                            'default_value'      => 0,
+                        ),
+                    );
+                    $available_options = array_merge( $available_options, $em_options );
 
-                    if ( class_exists( 'EM_Events' ) ) {
-                        $em_options        = array(
-                            array(
-                                'action_name'        => 'em_booking_delete',
-                                'action_generator'   => 'Events Manager',
-                                'action_title'       => 'Booking delete',
-                                'action_description' => esc_html( __( 'Logs when a booking is deleted', 'action-logger' ) ),
-                            ),
-                            array(
-                                'action_name'        => 'em_booking_cancel_reject',
-                                'action_generator'   => 'Events Manager',
-                                'action_title'       => 'Booking canceled/rejected',
-                                'action_description' => esc_html( __( 'Logs when a booking is canceled or rejected', 'action-logger' ) ),
-                            ),
-                        );
-                        $available_options = array_merge( $available_options, $em_options );
-                    }
+                    // csvi options
+                    $csvi_options      = array(
+                        array(
+                            'action_name'        => 'csvi_file_uploaded',
+                            'action_generator'   => 'CSV Importer',
+                            'action_title'       => 'CSV file uploaded',
+                            'action_description' => esc_html( __( 'Logs when a csv file is uploaded', 'action-logger' ) ),
+                            'default_value'      => 0,
+                        ),
+                    );
+                    $available_options = array_merge( $available_options, $csvi_options );
+    
+                    // add option for IDF rankings importer
+                    $ri_options      = array(
+                        array(
+                            'action_name'        => 'ri_file_uploaded',
+                            'action_generator'   => 'Rankings Importer',
+                            'action_title'       => 'CSV file uploaded',
+                            'action_description' => esc_html( __( 'Logs when a csv file is uploaded', 'action-logger' ) ),
+                            'default_value'      => 0,
+                        ),
+                    );
+                    $available_options = array_merge( $available_options, $ri_options );
+    
                     foreach ( $available_options as $option ) {
-                        update_option( 'al_' . $option[ 'action_name' ], true );
+                        update_option( 'al_' . $option[ 'action_name' ], $option[ 'default_value' ] );
                     }
                     update_option( 'al_available_log_actions', $available_options );
                     update_option( 'al_log_user_role', 'manage_options' );
                 }
             }
-
+    
             /**
              * Adds a page to admin sidebar menu
              */
@@ -263,7 +301,7 @@
                 add_menu_page( 'Action Logger', 'Action Logger', get_option( 'al_log_user_role' ), 'action-logger', 'action_logger_dashboard', 'dashicons-editor-alignleft' );
                 include( 'al-dashboard.php' ); // content for the settings page
             }
-
+    
             /**
              * Adds a (hidden) settings page, only through the menu on top of the pages.
              */
@@ -351,9 +389,7 @@
                             $get_available_actions = get_option( 'al_available_log_actions' );
                         }
                         foreach( $get_available_actions as $action ) {
-                            if ( ! isset( $_POST[ $action[ 'action_name' ] ] ) ) {
-                                delete_option( 'al_' . $action[ 'action_name' ] );
-                            } else {
+                            if ( isset( $_POST[ $action[ 'action_name' ] ] ) ) {
                                 update_option( 'al_' . $action[ 'action_name' ], 1 );
                             }
                         }
@@ -644,41 +680,57 @@
              * @param $booking_ids
              */
             public function al_log_registration_delete( $result, $booking_ids ) {
-                if ( class_exists( 'ActionLogger' ) && false != get_option( 'al_wp_em_booking_delete' ) ) {
+                if ( false != get_option( 'al_em_booking_delete' ) ) {
                     $this->al_log_user_action( 'registration_deleted', 'Action Logger', get_userdata( get_current_user_id() )->first_name . ' ' . esc_html( __( 'deleted bookings for an event.', 'action-logger' ) ) );
                 }
             }
-
+    
             /**
              * Log an action when an booking is canceled or rejected
              *
              * @param $EM_Event
              * @param $EM_Booking
              */
-            public function al_log_registration_cancel_reject( $EM_Event, $EM_Booking ) {
-
+            public function al_log_registration_chance( $EM_Event, $EM_Booking ) {
+    
+                // die('al_log_registration_chance');
                 if ( true == $EM_Event ) {
-
+    
+                    $show           = false;
                     $booking_id     = $EM_Booking->booking_id;
                     $booking_user   = $EM_Booking->person_id;
                     $booking_status = $EM_Booking->booking_status;
-                    if ( 2 == $EM_Booking->booking_status ) {
+    
+                    if ( 1 == $EM_Booking->booking_status ) {
+                        $action = 'registration_approved';
+                        $status = 'approved';
+                        if ( 1 == get_option( 'al_em_booking_approved' ) ) {
+                            $show = true;
+                        }
+                    } elseif ( 2 == $EM_Booking->booking_status ) {
                         $action = 'registration_reject';
                         $status = 'rejected';
+                        if ( 1 == get_option( 'al_em_booking_rejected' ) ) {
+                            $show = true;
+                        }
                     } elseif ( 3 == $EM_Booking->booking_status ) {
                         $action = 'registration_cancel';
                         $status = 'canceled';
+                        if ( 1 == get_option( 'al_em_booking_canceled' ) ) {
+                            $show = true;
+                        }
                     } else {
                         $action = 'registration_unknown';
-                        $status= 'unknown';
+                        $status = 'unknown';
+                        $show   = true;
                     }
-
-                    if ( class_exists( 'ActionLogger' ) && false != get_option( 'al_wp_em_booking_cancel_reject' ) ) {
+                    
+                    if ( $show == true ) {
                         $this->al_log_user_action( $action, 'Action Logger', get_userdata( get_current_user_id() )->first_name . ' ' . $status . ' ' . esc_html( __( 'the booking with booking ID:', 'action-logger' ) )  . ' ' . $booking_id . '.' );
                     }
                 }
             }
-
+    
         }
 
         /**
